backend/requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-jwt-extended
flask-cors
bcrypt
psycopg2-binary
python-dotenv
